refactor: update function prototypes and add tests for ft_substr, ft_strjoin, ft_strtrim, and ft_split

+++ libft/ft_strlen.c
@@ -6,11 +6,13 @@
 /*   By: jhoban <jhoban@student.42berlin.de>        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/10/19 16:55:42 by jhoban            #+#    #+#             */
-/*   Updated: 2025/11/19 14:07:16 by jhoban           ###   ########.fr       */
+/*   Updated: 2025/11/19 15:46:00 by jhoban           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-int	ft_strlen(const char *str)
+#include <string.h>
+
+size_t	ft_strlen(const char *str)
 {
 	int	index;
