+++ libft/ft_substr.c
@@ -9,7 +9,7 @@ char *ft_substr(char const *s, unsigned int start, size_t len)
 
     if (!s)
         return (NULL);
-    if (start >= ft_strlen(s))
+    if ((start >= (unsigned int)ft_strlen(s)))
         return (ft_strdup(""));
     // The length can't be greater than the remaining string length
     if (len > ft_strlen(s + start))
